+++ pkg/R/targetedLimma.R
@@ -1,34 +1,39 @@
-#' Variance Shrinkage of Linear Models for Microarrays on ATE parameter
-#' estimated via Targeted Minimum Loss-Based Estimation
+#' Moderated T-Statistic Test for Asymptotically Linear Target Parameters
 #'
 #' Performs variance shrinkage via empirical Bayes procedures on influence
 #' curve transforms of the Average Treatment Effect of gene expression
 #' changes on the outcome of interest
 #'
-#' @param ATE the average treatment effect of gene expression change on an
-#'            outcome of interest, as computed by \code{tmle.multiATE}
-#' @param Y the vector corresponding to the outcome of interest
+#' @param biomarkerTMLEout a generic object generated as output from the TMLE
+#'        procedure implemented in \code{biomarkerTMLE}
+#' @param IDs a vector providing the annotation information for the biomarkers,
+#'        genes, or other units of interested in the analysis
+#' @param designMat a design matrix providing the contrasts to be used in the
+#'        linear model fitting procedure of \code{limma::lmFit}
+#' @param ... additional arguments to be passed to functions from \code{limma}
 #'
-#' @importFrom limma limma
+#' @importFrom limma lmFit eBayes topTable
 #'
 #' @return
 #' A list containing two objects: (1) model fit returned by \code{limma::lmFit}
 #'                                (2) results table from \code{limma::topTable}
 #'
-#' @export tmle.limma
-#'
-#' @examples
+#' @export limmaTMLE
 #'
 
-tmle.limma <- function(ATE, Y) {
-  out <- vector("list", 2)
-  design <- as.data.frame(cbind(rep(1, nrow(Y)), Y$Y))
-  colnames(design) <- c("intercept", "Tx")
-  fit <- limma::lmFit(as.data.frame(ATE), design)
+limmaTMLE <- function(biomarkerTMLEout, IDs = NULL, designMat, ...) {
+
+  limmaTMLEout <- vector("list", 2)
+  fit <- limma::lmFit(as.data.frame(biomarkerTMLEout), designMat)
   fit <- limma::eBayes(fit)
   tt <- limma::topTable(fit, coef = 2, adjust.method = "BH", sorty.by = "none",
                         number = Inf)
-  out[[1]] <- fit
-  out[[2]] <- tt
-  return(out)
+  if (IDs != NULL) {
+    tt$IDs <- IDs
+  } else {
+    tt$IDs <- rownames(biomarkerTMLEout)
+  }
+  limmaTMLEout[[1]] <- fit
+  limmaTMLEout[[2]] <- tt
+  return(limmaTMLEout)
 }